+++ Plugins/UVOCBot.Plugins.Feeds/Extensions/IServiceCollectionExtensions.cs
@@ -1,5 +1,9 @@
-﻿using Microsoft.Extensions.DependencyInjection;
+﻿using Microsoft.Extensions.Configuration;
+using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.Options;
 using Remora.Commands.Extensions;
+using System;
+using Tweetinvi;
 using UVOCBot.Discord.Core.Extensions;
 using UVOCBot.Plugins.Feeds;
 using UVOCBot.Plugins.Feeds.Commands;
@@ -10,8 +14,12 @@ namespace UVOCBot.Plugins;
 
 public static class IServiceCollectionExtensions
 {
-    public static IServiceCollection AddFeedsPlugin(this IServiceCollection services)
+    public static IServiceCollection AddFeedsPlugin(this IServiceCollection services, IConfiguration config)
     {
+        services.Configure<FeedsPluginOptions>(config.GetSection(nameof(FeedsPluginOptions)));
+
+        services.AddTransient(TwitterClientFactory);
+
         services.AddComponentResponder<ToggleFeedComponentResponder>(FeedComponentKeys.ToggleFeed);
 
         services.AddCommandTree()
@@ -21,5 +29,12 @@ public static class IServiceCollectionExtensions
         services.AddHostedService<TwitterWorker>();
 
         return services;
+    }
+
+    private static ITwitterClient TwitterClientFactory(IServiceProvider services)
+    {
+        FeedsPluginOptions options = services.GetRequiredService<IOptions<FeedsPluginOptions>>().Value;
+
+        return new TwitterClient(options.TwitterKey, options.TwitterSecret, options.TwitterBearerToken);
     }
 }